Download charts.
It is now possible to download charts from the services (right now a local node server). The zip file is downloaded, unzipped, the chart images are then saved to the file system, and the zip is deleted. Since this work is asynchronous, we must be careful not to set the state of an unmounted component. If an error occurs, appropriate messages are displayed.

This commit also add react-native-fs (https://github.com/johanneslumpe/react-native-fs) and react-native-progress (https://github.com/oblador/react-native-progress).